Move extensions and helpers to own files
