+++ Sources/RemoteImage/RemoteImage.swift
@@ -88,12 +88,12 @@ public struct RemoteImage<Content: View>: View {
     configuration: RemoteImageConfiguration = .init(),
     @ViewBuilder content: @escaping (RemoteImagePhase) -> Content)
   {
-    _model = .init(
-      wrappedValue: RemoteImageViewModel(
+    self.init(
+      model: RemoteImageViewModel(
         url: url,
         urlSession: urlSession,
-        configuration: configuration))
-    self.content = content
+        configuration: configuration),
+      content: content)
   }
 
   /// Initialize a new `RemoteImage` instance using either the fetched remote image or an empty fallback.
@@ -162,6 +162,14 @@ public struct RemoteImage<Content: View>: View {
 
   // MARK: Internal
 
+  init(
+    model: RemoteImageViewModel,
+    @ViewBuilder content: @escaping (RemoteImagePhase) -> Content)
+  {
+    _model = .init(wrappedValue: model)
+    self.content = content
+  }
+
   @StateObject var model: RemoteImageViewModel
 
   var didAppear: ((Self) -> Void)?
@@ -170,118 +178,6 @@ public struct RemoteImage<Content: View>: View {
 
   private var content: (RemoteImagePhase) -> Content
 
-  @ViewBuilder
-  private static func imageForPhaseOrEmpty(
-    _ phase: RemoteImagePhase)
-    -> Content
-    where
-    Content == _ConditionalContent<Image, Image>
-  {
-    if let image = phase.image {
-      image
-    } else {
-      Image(nativeImage: .init())
-    }
-  }
-
-  @ViewBuilder
-  private static func imageForPhaseOrPlaceholder<I, P>(
-    _ phase: RemoteImagePhase,
-    @ViewBuilder content: @escaping (Image) -> I,
-    @ViewBuilder placeholder: @escaping () -> P)
-    -> Content
-    where
-    Content == _ConditionalContent<I, P>,
-    I: View,
-    P: View
-  {
-    if let image = phase.image {
-      content(image)
-    } else {
-      placeholder()
-    }
-  }
-}
-
-// MARK: initWithCache
-
-extension RemoteImage {
-
-  /// Initialize a new `RemoteImage` instance.
-  ///
-  /// A ``URLSession`` will be constructed using the ``URLSessionConfiguration.default`` configuration and the specified ``cache``.
-  /// - Parameters:
-  ///   - url: The URL of the image to display.
-  ///   - cache: Cache to use with the underlying ``URLSession``.
-  ///   - configuration: Configuration options to use. If none is provided, defaults values are used. See ``RemoteImageConfiguration``.
-  ///   - content: A closure that takes the load phase as an input, and returns the view to display for the specified phase.
-  public init(
-    url: URL?,
-    cache: URLCache,
-    configuration: RemoteImageConfiguration = .init(),
-    @ViewBuilder content: @escaping (RemoteImagePhase) -> Content)
-  {
-    _model = .init(
-      wrappedValue: RemoteImageViewModel(
-        url: url,
-        cache: cache,
-        configuration: configuration))
-    self.content = content
-  }
-
-  /// Initialize a new `RemoteImage` instance using either the fetched remote image or an empty fallback.
-  ///
-  /// A ``URLSession`` will be constructed using the ``URLSessionConfiguration.default`` configuration and the specified ``cache``.
-  /// - Parameters:
-  ///   - url: The URL of the image to display.
-  ///   - cache: Cache to use with the underlying ``URLSession``.
-  ///   - configuration: Configuration options to use. If none is provided, defaults values are used. See ``RemoteImageConfiguration``.
-  public init(
-    url: URL?,
-    cache: URLCache,
-    configuration: RemoteImageConfiguration = .init())
-    where
-    Content == _ConditionalContent<Image, Image>
-  {
-    self.init(
-      url: url,
-      cache: cache,
-      configuration: configuration,
-      content: Self.imageForPhaseOrEmpty)
-  }
-
-  /// Initialize a new `RemoteImage` instance using a custom placeholder.
-  ///
-  /// A ``URLSession`` will be constructed using the ``URLSessionConfiguration.default`` configuration and the specified ``cache``.
-  /// - Parameters:
-  ///   - url: The URL of the image to display.
-  ///   - cache: Cache to use with the underlying ``URLSession``.
-  ///   - configuration: Configuration options to use. If none is provided, defaults values are used. See ``RemoteImageConfiguration``.
-  ///   - content: A closure that takes the loaded image as an input, and returns the view to show. You can return the image directly, or modify it as needed
-  ///     before returning it.
-  ///   - placeholder: A closure that returns the view to show until the load operation completes successfully.
-  public init<I, P>(
-    url: URL?,
-    cache: URLCache,
-    configuration: RemoteImageConfiguration = .init(),
-    @ViewBuilder content: @escaping (Image) -> I,
-    @ViewBuilder placeholder: @escaping () -> P)
-    where
-    Content == _ConditionalContent<I, P>,
-    I: View,
-    P: View
-  {
-    self.init(
-      url: url,
-      cache: cache,
-      configuration: configuration)
-    { phase in
-      Self.imageForPhaseOrPlaceholder(
-        phase,
-        content: content,
-        placeholder: placeholder)
-    }
-  }
 }
 
 #if DEBUG
